requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask==2.3.3
Flask-CORS==4.0.0
PyDrive==1.3.1
google-api-python-client==2.103.0
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
numpy>=1.24.0
scipy>=1.10.0
torch==2.0.1
torchvision==0.15.2
efficientnet-pytorch==0.7.1
Pillow>=9.5.0
